+++ modelcraft/environ.py
@@ -6,7 +6,6 @@ def setup_environ():
         if variable not in os.environ:
             raise EnvironmentError(variable + " environment variable not set")
     os.environ["LD_LIBRARY_PATH"] = os.environ["CLIB"]
-    os.environ["COOT_N_THREADS"] = "1"
     os.environ["GOTO_NUM_THREADS"] = "1"
     os.environ["MKL_NUM_THREADS"] = "1"
     os.environ["NUMEXPR_NUM_THREADS"] = "1"