+++ modelcraft/tests/ccp4/test_prune.py
@@ -1,16 +1,17 @@
-from modelcraft.jobs.coot import Prune
-from modelcraft.structure import ModelStats
+from ...structure import ModelStats
+from ...prune import prune
 from . import insulin_refmac
 
 
 def test_insulin_prune():
     refmac = insulin_refmac()
-    refmac.structure.remove_alternative_conformations()
-    coot = Prune(
+    pruned = prune(
         structure=refmac.structure,
         fphi_best=refmac.fphi_best,
         fphi_diff=refmac.fphi_diff,
-    ).run()
+        fphi_calc=refmac.fphi_calc,
+        threshold=-2,
+    )
     stats_in = ModelStats(refmac.structure)
-    stats_out = ModelStats(coot.structure)
+    stats_out = ModelStats(pruned)
     assert stats_out.residues < stats_in.residues